requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
lyzr>=0.1.0
pymongo==4.9.0
requests==2.31.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
twilio==9.3.0
httpx==0.27.2
websockets==14.1
redis>=5.0.0
PyJWT>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

